refactor: replace types
