+++ src/components/cardText/index.tsx
@@ -1,9 +1,6 @@
 import { FC } from 'react';
 import { StyledCard } from './syled';
-
-export interface CardtextProps {
-    children: React.ReactNode;
-}
+import { CardtextProps } from './types';
 
 export const CardText: FC<CardtextProps> = ({ children }) => {
     return <StyledCard>{children}</StyledCard>;
