+++ src/components/cardText/types.ts
@@ -1,0 +1,3 @@
+export interface CardtextProps {
+    children: React.ReactNode;
+}
