+++ src/components/meny/menyButton/styled.ts
@@ -10,11 +10,11 @@ export const StyledButton = styled('button')`
     border-bottom: 20px solid white;
     border-left: 19px solid transparent;
     border-right: 19px solid transparent;
-
     background-color: transparent;
     padding: 0;
     border-radius: 0;
     rotate: 90deg;
+
     svg {
         path {
             fill: #8c8c8c;
@@ -24,9 +24,11 @@ export const StyledButton = styled('button')`
     &:hover {
         color: none;
     }
+
     &:focus {
         outline: 0;
     }
+
     @media screen and (max-width: 730px) {
         display: none;
     }
