+++ src/components/meny/menyButton/types.ts
@@ -1,0 +1,4 @@
+export interface ButtonMenuProps {
+    collapsed: boolean;
+    onClick: () => void;
+}
